mtmd : move helpers to dedicated file
